+++ hmf/__init__.py
@@ -1,6 +1,13 @@
 __version__ = "3.0.2"
-from .hmf import MassFunction
-from . import fitting_functions as fits
-from .cosmo import Cosmology
-from .transfer import Transfer
-from .sample import sample_mf
+
+from .mass_function.hmf import MassFunction
+
+# To patch the transition to modularised form, import modules here.
+# Perhaps I should deprecate this for some version.
+from .alternatives import *
+from .cosmology import *
+from .density_field import *
+from .fitting import *
+from .halos import *
+from .helpers import *
+from .mass_function import *