+++ hmf/cosmology/__init__.py
@@ -1,0 +1,3 @@
+"""
+A subpackage dedicated to cosmographic calculations, and other purely cosmological quantities, such as the growth factor.
+"""