+++ hmf/cosmology/cosmo.py
@@ -11,9 +11,8 @@ Cosmology framework. All custom subclasses of :class:`astropy.cosmology.FLRW`
 may be used as inputs.
 """
 
-from . import _cache
-from astropy.cosmology import Planck13, FLRW, WMAP5, WMAP7, WMAP9, Planck15
-from . import _framework
+from astropy.cosmology import FLRW, Planck15, WMAP5, WMAP7, WMAP9, Planck13
+from .._internals import _framework, _cache
 import sys
 import astropy.units as u