+++ hmf/density_field/__init__.py
@@ -1,0 +1,4 @@
+"""
+A subpackage dedicated to basic measures of the matter density field -- its 2-point structure, cosmological transfer
+functions, and filter functions which can be applied to it.
+"""