+++ hmf/helpers/sample.py
@@ -4,7 +4,7 @@ Module for dealing with sampled mass functions.
 Provides routines for sampling theoretical functions, and for binning sampled data.
 '''
 import numpy as np
-from . import hmf
+from ..mass_function import hmf
 from scipy.interpolate import InterpolatedUnivariateSpline as _spline